fix(utxos): remove spent utxos

+++ bin/btc-server/src/bin/main.rs
@@ -1144,6 +1144,7 @@ where
                     }
                     _ => {
                         error!("Failed to broadcast transaction: {}", err_msg);
+                        error!("Failed tx: {:?}", tx);
                         Err(CoordinatorError::FailedToBroadcastTx(err))
                     }
                 }
@@ -1571,19 +1572,38 @@ where
     }
 }
 
-impl<BitcoindClient> App<BitcoindClient> {
+impl<BitcoindClient: bitcoincore_rpc::RpcApi> App<BitcoindClient> {
+    /// Handles invalid inputs in the transaction by:
+    /// 1. Checking if the input's previous output exists in the database.
+    /// 2. If it exists, checks if it's already spent.
+    /// 3. If it is spent, removes it from the database.
+    ///
+    /// Returns `Ok(())` if all inputs are handled successfully, or an error if any operation fails.
     fn handle_invalid_inputs(&self, tx: &Transaction) -> Result<(), btcserverlib::database::Error> {
+        let tx_id = tx.compute_txid();
         for input in &tx.input {
             if let Some(_utxo) = self.db.get_utxo(input.previous_output)? {
-                match self.db.remove_utxo(&input.previous_output) {
-                    Ok(_) => {
-                        info!("Removed spent input: {} from DB", input.previous_output);
-                    }
-                    Err(e) => {
-                        error!(
-                            "Failed to remove spent input: {} from DB: {}",
-                            input.previous_output, e
-                        );
+                // Check on chain if the input is already spent
+                let result = self
+                    .bitcoind_client
+                    .get_tx_out(&tx_id, input.previous_output.vout, None)
+                    .map_err(|e| {
+                        error!("Failed to get tx out for input: {}: {}", input.previous_output, e);
+                        btcserverlib::database::Error::BitcoindError(e)
+                    })?;
+
+                if result.is_none() {
+                    // The input is already spent, remove it from the database
+                    match self.db.remove_utxo(&input.previous_output) {
+                        Ok(_) => {
+                            info!("Removed spent input: {} from DB", input.previous_output);
+                        }
+                        Err(e) => {
+                            error!(
+                                "Failed to remove spent input: {} from DB: {}",
+                                input.previous_output, e
+                            );
+                        }
                     }
                 }
             }
